f_pc_draw_priority: Type fix

+++ include/f_pc/f_pc_draw_priority.h
@@ -8,7 +8,7 @@ typedef struct draw_priority_class {
     s16 mPriority;
 } draw_priority_class;
 
-s32 fpcDwPi_Get(const draw_priority_class* pDwPi);
+s16 fpcDwPi_Get(const draw_priority_class* pDwPi);
 void fpcDwPi_Set(draw_priority_class* pDwPi, s16 p);
 void fpcDwPi_Init(draw_priority_class* pDwPi, s16 p);
